feat" rename icon to headericon
